chore(langgraph): simplify MultiAbortSignal interface
We can't really use class inheritance to tack on dependent signals because we can't naturally construct AbortSignal. We can make the class take on the same interface as MDN AbortSignal, but if it's missing special node-only properties then it causes unexpected behavior.

This simplifies the pattern to only use type guards and interfaces to track dependent signals.